+++ age-plugin/src/recipient.rs
@@ -2,12 +2,16 @@
 
 use age_core::{
     format::{FileKey, Stanza, FILE_KEY_BYTES},
-    plugin::{Connection, UnidirSend},
+    plugin::{self, BidirSend, Connection},
 };
+use secrecy::SecretString;
 use std::convert::TryInto;
 use std::io;
 
+use crate::Callbacks;
+
 const ADD_RECIPIENT: &str = "add-recipient";
+const ADD_IDENTITY: &str = "add-identity";
 const WRAP_FILE_KEY: &str = "wrap-file-key";
 const RECIPIENT_STANZA: &str = "recipient-stanza";
 
@@ -24,11 +28,76 @@ pub trait RecipientPluginV1 {
         recipients: I,
     ) -> Result<(), Vec<Error>>;
 
-    /// Wraps `file_key` to all recipients previously added via `add_recipients`.
+    /// Stores identities that the user would like to encrypt age files to.
     ///
-    /// Returns either one stanza per recipient, or any errors if one or more recipients
-    /// could not be wrapped to.
-    fn wrap_file_key(&mut self, file_key: &FileKey) -> Result<Vec<Stanza>, Vec<Error>>;
+    /// Each identity string is Bech32-encoded with an HRP of `AGE-PLUGIN-NAME-` where
+    /// `NAME` is the name of the plugin that resolved to this binary.
+    ///
+    /// Returns a list of errors if any of the identities are unknown or invalid.
+    fn add_identities<'a, I: Iterator<Item = &'a str>>(
+        &mut self,
+        identities: I,
+    ) -> Result<(), Vec<Error>>;
+
+    /// Wraps each `file_key` to all recipients and identities previously added via
+    /// `add_recipients` and `add_identities`.
+    ///
+    /// Returns either one stanza per recipient and identity for each file key, or any
+    /// errors if one or more recipients or identities could not be wrapped to.
+    ///
+    /// `callbacks` can be used to interact with the user, to have them take some physical
+    /// action or request a secret value.
+    fn wrap_file_keys(
+        &mut self,
+        file_keys: Vec<FileKey>,
+        callbacks: impl Callbacks<Error>,
+    ) -> io::Result<Result<Vec<Vec<Stanza>>, Vec<Error>>>;
+}
+
+/// The interface that age plugins can use to interact with an age implementation.
+struct BidirCallbacks<'a, 'b, R: io::Read, W: io::Write>(&'b mut BidirSend<'a, R, W>);
+
+impl<'a, 'b, R: io::Read, W: io::Write> Callbacks<Error> for BidirCallbacks<'a, 'b, R, W> {
+    /// Shows a message to the user.
+    ///
+    /// This can be used to prompt the user to take some physical action, such as
+    /// inserting a hardware key.
+    fn message(&mut self, message: &str) -> plugin::Result<(), ()> {
+        self.0
+            .send("msg", &[], message.as_bytes())
+            .map(|res| res.map(|_| ()))
+    }
+
+    fn request_public(&mut self, message: &str) -> plugin::Result<String, ()> {
+        self.0
+            .send("request-public", &[], message.as_bytes())
+            .and_then(|res| match res {
+                Ok(s) => String::from_utf8(s.body)
+                    .map_err(|_| {
+                        io::Error::new(io::ErrorKind::InvalidData, "response is not UTF-8")
+                    })
+                    .map(Ok),
+                Err(()) => Ok(Err(())),
+            })
+    }
+
+    /// Requests a secret value from the user, such as a passphrase.
+    ///
+    /// `message` will be displayed to the user, providing context for the request.
+    fn request_secret(&mut self, message: &str) -> plugin::Result<SecretString, ()> {
+        self.0
+            .send("request-secret", &[], message.as_bytes())
+            .and_then(|res| match res {
+                Ok(s) => String::from_utf8(s.body)
+                    .map_err(|_| io::Error::new(io::ErrorKind::InvalidData, "secret is not UTF-8"))
+                    .map(|s| Ok(SecretString::new(s))),
+                Err(()) => Ok(Err(())),
+            })
+    }
+
+    fn error(&mut self, error: Error) -> plugin::Result<(), ()> {
+        error.send(self.0).map(|()| Ok(()))
+    }
 }
 
 /// The kinds of errors that can occur within the recipient plugin state machine.
@@ -36,6 +105,13 @@ pub enum Error {
     /// An error caused by a specific recipient.
     Recipient {
         /// The index of the recipient.
+        index: usize,
+        /// The error message.
+        message: String,
+    },
+    /// An error caused by a specific identity.
+    Identity {
+        /// The index of the identity.
         index: usize,
         /// The error message.
         message: String,
@@ -51,6 +127,7 @@ impl Error {
     fn kind(&self) -> &str {
         match self {
             Error::Recipient { .. } => "recipient",
+            Error::Identity { .. } => "identity",
             Error::Internal { .. } => "internal",
         }
     }
@@ -58,13 +135,16 @@ impl Error {
     fn message(&self) -> &str {
         match self {
             Error::Recipient { message, .. } => &message,
+            Error::Identity { message, .. } => &message,
             Error::Internal { message } => &message,
         }
     }
 
-    fn send<R: io::Read, W: io::Write>(self, phase: &mut UnidirSend<R, W>) -> io::Result<()> {
+    fn send<R: io::Read, W: io::Write>(self, phase: &mut BidirSend<R, W>) -> io::Result<()> {
         let index = match self {
-            Error::Recipient { index, .. } => Some(index.to_string()),
+            Error::Recipient { index, .. } | Error::Identity { index, .. } => {
+                Some(index.to_string())
+            }
             Error::Internal { .. } => None,
         };
 
@@ -73,7 +153,11 @@ impl Error {
             None => vec![self.kind()],
         };
 
-        phase.send("error", &metadata, self.message().as_bytes())
+        phase
+            .send("error", &metadata, self.message().as_bytes())?
+            .unwrap();
+
+        Ok(())
     }
 }
 
@@ -82,8 +166,8 @@ pub(crate) fn run_v1<P: RecipientPluginV1>(mut plugin: P) -> io::Result<()> {
     let mut conn = Connection::accept();
 
     // Phase 1: collect recipients, and file keys to be wrapped
-    let (recipients, file_keys) = {
-        let (recipients, file_keys) = conn.unidir_receive(
+    let ((recipients, identities), file_keys) = {
+        let (recipients, identities, file_keys) = conn.unidir_receive(
             (ADD_RECIPIENT, |s| {
                 if s.args.len() == 1 && s.body.is_empty() {
                     Ok(s)
@@ -96,7 +180,19 @@ pub(crate) fn run_v1<P: RecipientPluginV1>(mut plugin: P) -> io::Result<()> {
                     })
                 }
             }),
-            (WRAP_FILE_KEY, |s| {
+            (ADD_IDENTITY, |s| {
+                if s.args.len() == 1 && s.body.is_empty() {
+                    Ok(s)
+                } else {
+                    Err(Error::Internal {
+                        message: format!(
+                            "{} command must have exactly one metadata argument and no data",
+                            ADD_IDENTITY
+                        ),
+                    })
+                }
+            }),
+            (Some(WRAP_FILE_KEY), |s| {
                 // TODO: Should we ignore file key commands with unexpected metadata args?
                 TryInto::<[u8; FILE_KEY_BYTES]>::try_into(&s.body[..])
                     .map_err(|_| Error::Internal {
@@ -106,13 +202,19 @@ pub(crate) fn run_v1<P: RecipientPluginV1>(mut plugin: P) -> io::Result<()> {
             }),
         )?;
         (
-            match recipients {
-                Ok(r) if r.is_empty() => Err(vec![Error::Internal {
-                    message: format!("Need at least one {} command", ADD_RECIPIENT),
-                }]),
+            match (recipients, identities) {
+                (Ok(r), Ok(i)) if r.is_empty() && i.is_empty() => (
+                    Err(vec![Error::Internal {
+                        message: format!(
+                            "Need at least one {} or {} command",
+                            ADD_RECIPIENT, ADD_IDENTITY
+                        ),
+                    }]),
+                    Err(vec![]),
+                ),
                 r => r,
             },
-            match file_keys {
+            match file_keys.unwrap() {
                 Ok(f) if f.is_empty() => Err(vec![Error::Internal {
                     message: format!("Need at least one {} command", WRAP_FILE_KEY),
                 }]),
@@ -122,17 +224,17 @@ pub(crate) fn run_v1<P: RecipientPluginV1>(mut plugin: P) -> io::Result<()> {
     };
 
     // Phase 2: wrap the file keys or return errors
-    conn.unidir_send(|mut phase| {
-        let (recipients, file_keys) = match (recipients, file_keys) {
-            (Ok(recipients), Ok(file_keys)) => (recipients, file_keys),
-            (Err(errors1), Err(errors2)) => {
-                for error in errors1.into_iter().chain(errors2.into_iter()) {
-                    error.send(&mut phase)?;
-                }
-                return Ok(());
-            }
-            (Err(errors), _) | (_, Err(errors)) => {
-                for error in errors {
+    conn.bidir_send(|mut phase| {
+        let (recipients, identities, file_keys) = match (recipients, identities, file_keys) {
+            (Ok(recipients), Ok(identities), Ok(file_keys)) => (recipients, identities, file_keys),
+            (recipients, identities, file_keys) => {
+                for error in recipients
+                    .err()
+                    .into_iter()
+                    .chain(identities.err())
+                    .chain(file_keys.err())
+                    .flatten()
+                {
                     error.send(&mut phase)?;
                 }
                 return Ok(());
@@ -145,25 +247,25 @@ pub(crate) fn run_v1<P: RecipientPluginV1>(mut plugin: P) -> io::Result<()> {
             for error in errors {
                 error.send(&mut phase)?;
             }
+        } else if let Err(errors) =
+            plugin.add_identities(identities.iter().map(|s| s.args.first().unwrap().as_str()))
+        {
+            for error in errors {
+                error.send(&mut phase)?;
+            }
         } else {
-            match file_keys
-                .into_iter()
-                .map(|file_key| plugin.wrap_file_key(&file_key))
-                .collect::<Result<Vec<_>, _>>()
-            {
+            match plugin.wrap_file_keys(file_keys, BidirCallbacks(&mut phase))? {
                 Ok(files) => {
                     for (file_index, stanzas) in files.into_iter().enumerate() {
-                        // The plugin MUST generate an error if one or more
-                        // recipients cannot be wrapped to. And it's a programming
-                        // error to return more stanzas than recipients.
-                        assert_eq!(stanzas.len(), recipients.len());
+                        // The plugin MUST generate an error if one or more recipients or
+                        // identities cannot be wrapped to. And it's a programming error
+                        // to return more stanzas than recipients and identities.
+                        assert_eq!(stanzas.len(), recipients.len() + identities.len());
 
                         for stanza in stanzas {
-                            phase.send_stanza(
-                                RECIPIENT_STANZA,
-                                &[&file_index.to_string()],
-                                &stanza,
-                            )?;
+                            phase
+                                .send_stanza(RECIPIENT_STANZA, &[&file_index.to_string()], &stanza)?
+                                .unwrap();
                         }
                     }
                 }